update docs and add headers, also ignore some files
