+++ Makefile
@@ -1,13 +1,20 @@
 CC=gcc
 CFLAGS=-Wall -Wextra -Werror
+OBJS = editor.o common.o console.o
 
 all: editor
 
-editor: editor.c editor.h
-	$(CC) $(CFLAGS) -O2 -o editor $<
+editor: clean_objs $(OBJS)
+	$(CC) $(CFLAGS) -o editor $(OBJS)
 
-debug: editor.c editor.h
-	$(CC) $(CFLAGS) -g -O0 -o editor $<
+%.o: %.c
+	$(CC) $(CFLAGS) -c $<
+
+clean_objs:
+	rm -f $(OBJS)
 
 clean:
-	rm -f editor
+	rm -f editor $(OBJS)
+
+debug: clean_objs $(OBJS)
+	$(CC) $(CFLAGS) -g -O0 -o editor $(OBJS)
