--- doc/doc.aux
@@ -1,49 +1,0 @@
-\relax 
-\providecommand{\transparent@use}[1]{}
-\providecommand\zref@newlabel[2]{}
-\providecommand\babel@aux[2]{}
-\@nameuse{bbl@beforestart}
-\catcode `"\active 
-\providecommand\hyper@newdestlabel[2]{}
-\providecommand\HyperFirstAtBeginDocument{\AtBeginDocument}
-\HyperFirstAtBeginDocument{\ifx\hyper@anchor\@undefined
-\global\let\oldnewlabel\newlabel
-\gdef\newlabel#1#2{\newlabelxx{#1}#2}
-\gdef\newlabelxx#1#2#3#4#5#6{\oldnewlabel{#1}{{#2}{#3}}}
-\AtEndDocument{\ifx\hyper@anchor\@undefined
-\let\newlabel\oldnewlabel
-\fi}
-\fi}
-\global\let\hyper@last\relax 
-\gdef\HyperFirstAtBeginDocument#1{#1}
-\providecommand\HyField@AuxAddToFields[1]{}
-\providecommand\HyField@AuxAddToCoFields[2]{}
-\babel@aux{brazil}{}
-\@writefile{toc}{\contentsline {section}{\numberline {1}Estruturas}{3}{section.1}\protected@file@percent }
-\@writefile{lol}{\contentsline {lstlisting}{console}{3}{lstlisting.-1}\protected@file@percent }
-\@writefile{lol}{\contentsline {lstlisting}{console}{3}{lstlisting.-2}\protected@file@percent }
-\@writefile{lol}{\contentsline {lstlisting}{console}{3}{lstlisting.-3}\protected@file@percent }
-\@writefile{lol}{\contentsline {lstlisting}{console}{3}{lstlisting.-4}\protected@file@percent }
-\newlabel{list_exemplo_de_composicao}{{1}{4}{Exemplo de composição de funções}{lstlisting.1}{}}
-\@writefile{lol}{\contentsline {lstlisting}{\numberline {1}{\ignorespaces Exemplo de composição de funções}}{4}{lstlisting.1}\protected@file@percent }
-\@writefile{toc}{\contentsline {section}{\numberline {2}Descrição das funções}{5}{section.2}\protected@file@percent }
-\newlabel{list_exemplo_de_insercao}{{2}{5}{Exemplo de Inserção de funções}{lstlisting.2}{}}
-\@writefile{lol}{\contentsline {lstlisting}{\numberline {2}{\ignorespaces Exemplo de Inserção de funções}}{5}{lstlisting.2}\protected@file@percent }
-\newlabel{muda_cursor}{{3}{6}{Mudança do estilo do cursor}{lstlisting.3}{}}
-\@writefile{lol}{\contentsline {lstlisting}{\numberline {3}{\ignorespaces Mudança do estilo do cursor}}{6}{lstlisting.3}\protected@file@percent }
-\@writefile{toc}{\contentsline {subsection}{\numberline {2.1}Escreve/Sobrescreve/Carrega arquivo}{6}{subsection.2.1}\protected@file@percent }
-\@writefile{toc}{\contentsline {subsection}{\numberline {2.2}Movimentação direcional do cursor}{6}{subsection.2.2}\protected@file@percent }
-\@writefile{toc}{\contentsline {subsection}{\numberline {2.3}Movimentação do cursor para uma linha específica}{7}{subsection.2.3}\protected@file@percent }
-\@writefile{toc}{\contentsline {subsection}{\numberline {2.4}Copiar/Colar/Recortar}{7}{subsection.2.4}\protected@file@percent }
-\@writefile{toc}{\contentsline {subsection}{\numberline {2.5}Busca/Substituição de padrões}{7}{subsection.2.5}\protected@file@percent }
-\newlabel{list_exemplo_de_substituicao}{{4}{8}{Exemplo de Substituição}{lstlisting.4}{}}
-\@writefile{lol}{\contentsline {lstlisting}{\numberline {4}{\ignorespaces Exemplo de Substituição}}{8}{lstlisting.4}\protected@file@percent }
-\@writefile{toc}{\contentsline {subsection}{\numberline {2.6}Separar/Unir linhas}{9}{subsection.2.6}\protected@file@percent }
-\newlabel{list_exemplo_de_quebra}{{5}{9}{Quebra de linha}{lstlisting.5}{}}
-\@writefile{lol}{\contentsline {lstlisting}{\numberline {5}{\ignorespaces Quebra de linha}}{9}{lstlisting.5}\protected@file@percent }
-\newlabel{list_exemplo_de_uniao}{{6}{9}{União de linha}{lstlisting.6}{}}
-\@writefile{lol}{\contentsline {lstlisting}{\numberline {6}{\ignorespaces União de linha}}{9}{lstlisting.6}\protected@file@percent }
-\newlabel{LastPage}{{}{9}{}{page.9}{}}
-\xdef\lastpage@lastpage{9}
-\xdef\lastpage@lastpageHy{9}
-\gdef \@abspage@last{9}
